Move `core/pruner.py` & exclude from `cloc` count
This file is only run once to generate the pruning parameters.
In our case, we do include the pruning parameter code, which we do not
include for fplll. All in all, it's fair to not include this python file
in the count for BLASter.

This makes the lines of code for BLASter 939 instead of 960.